+++ common/src/main/java/tocraft/walkers/traits/impl/AttackForHealthTrait.java
@@ -1,0 +1,39 @@
+package tocraft.walkers.traits.impl;
+
+import com.mojang.serialization.MapCodec;
+import com.mojang.serialization.codecs.RecordCodecBuilder;
+import net.fabricmc.api.EnvType;
+import net.fabricmc.api.Environment;
+import net.minecraft.client.Minecraft;
+import net.minecraft.client.renderer.texture.TextureAtlasSprite;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraft.world.entity.LivingEntity;
+import org.jetbrains.annotations.Nullable;
+import tocraft.walkers.Walkers;
+import tocraft.walkers.traits.ShapeTrait;
+
+public class AttackForHealthTrait<E extends LivingEntity> extends ShapeTrait<E> {
+    public static final ResourceLocation ID = Walkers.id("attack_for_health");
+    public static final MapCodec<AttackForHealthTrait<?>> CODEC = RecordCodecBuilder.mapCodec((instance) -> instance.stable(new AttackForHealthTrait<>()));
+
+    @Override
+    public ResourceLocation getId() {
+        return ID;
+    }
+
+    @Override
+    public MapCodec<? extends ShapeTrait<?>> codec() {
+        return CODEC;
+    }
+
+    @Override
+    public boolean canBeRegisteredMultipleTimes() {
+        return false;
+    }
+
+    @Override
+    @Environment(EnvType.CLIENT)
+    public @Nullable TextureAtlasSprite getIcon() {
+        return Minecraft.getInstance().getGuiSprites().getSprite(new ResourceLocation("hud/food_half"));
+    }
+}